phase-3/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
mcp>=1.0.0
groq>=0.4.0
pydantic==2.5.3
